+++ src/eventbus.js
@@ -1,3 +1,3 @@
 import Vue from 'vue';
-export const eventbus=new Vue();
+export const eventbus = new Vue();
 //이벤트 버스(컴포넌트간 데이터 전달)를 사용하기 위한 파일
