+++ src/views/Home.vue
@@ -12,8 +12,8 @@
 <script>
 
 import Intro from '@/components/Home/Intro.vue'
-import FilterSection from '@/components/Home/Filter/FilterSection.vue'
-import Gallery from '@/components/Gallery/MainGallery/Gallery.vue'
+import FilterSection from '@/components/Home/FilterSection.vue'
+import Gallery from '@/components/Gallery/Gallery.vue'
 
 import img1 from '@/assets/img1.jpg'
 import img2 from '@/assets/img2.jpg'
@@ -75,8 +75,6 @@ export default {
         show: false
       },
     ];
-
-
     return {
       myImages,
       file:'',
